CLAS-365: Move to contractTest package
