+++ claim-for-payment-service/src/contractTest/java/uk/gov/justice/laa/claimforpayment/config/ClaimsApiPactTestConfig.java
@@ -1,4 +1,4 @@
-package contract.config;
+package uk.gov.justice.laa.claimforpayment.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
